% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{addition}
\alias{addition}
\title{Basic addition method}
\usage{
addition(a, b)
}
\arguments{
\item{a}{First parameter}

\item{b}{Second parameter}
}
\value{
the result
}
\description{
A sample method which can be used to demonstrate a real package.
}
