% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{square}
\alias{square}
\title{Square, using a method in C}
\usage{
square(x)
}
\arguments{
\item{x}{The parameter to square}
}
\value{
the result
}
\description{
A sample method which calls a function in C
}
