% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{subtraction}
\alias{subtraction}
\title{Basic subtraction method}
\usage{
subtraction(a, b)
}
\arguments{
\item{a}{First parameter}

\item{b}{Second parameter}
}
\value{
the result
}
\description{
A sample method which can be used to demonstrate a real package.
}
